Start fiddling with actual matrix
We now have a working POC. Time to start writing the actual matrix
scanning and layout. We need to list all the rows and the cols, and then
figure out how to map the pressed positions into keys. Thinking about
running a task for scanning and then using `Shared` to keep the state of
what has been pressed. Then another task to map that to keycodes and
then another task to send the datagram. We'll see how that plans out.