Scanning works for one dir
Using the neopixel as debug (awesome stuff that one) I can now show that
scanning in one direction works. It also shows that state changes kinda
works. For some reason it does not seem to register when two keys in the
same column are activated. Will need to debug that.

Next up is scanning in the other direction as well. And after that why
it does not register when two keys in the same column are pressed.